resolve unit test

+++ features/content/play/src/test/java/com/tokopedia/play/viewmodel/play/PlayViewModelFieldTest.kt
@@ -267,7 +267,7 @@ class PlayViewModelFieldTest {
                                     likeInfo = likeBuilder.buildCompleteData(
                                             param = initialChannelData.likeInfo.param,
                                             status = likeBuilder.buildStatus(
-                                                    totalLike = totalLike.toInt(),
+                                                    totalLike = totalLike.toLong(),
                                                     totalLikeFormatted = totalLike,
                                                     isLiked = isLiked,
                                                     source = LikeSource.Network
