+++ features/content/play/src/test/java/com/tokopedia/play/viewmodel/play/PlayViewModelLikeTest.kt
@@ -64,7 +64,7 @@ class PlayViewModelLikeTest {
                 likeInfo = likeModelBuilder.buildCompleteData(
                         param = likeModelBuilder.buildParam(),
                         status = likeModelBuilder.buildStatus(
-                                totalLike = initialTotalLike,
+                                totalLike = initialTotalLike.toLong(),
                                 totalLikeFormatted = initialTotalLike.toString(),
                                 isLiked = false
                         )
@@ -82,7 +82,7 @@ class PlayViewModelLikeTest {
                     .isAvailable()
                     .isEqualTo(
                             likeModelBuilder.buildStatus(
-                                    totalLike = newLikeCount,
+                                    totalLike = newLikeCount.toLong(),
                                     totalLikeFormatted = newLikeCount.toString(),
                                     isLiked = true,
                                     source = LikeSource.UserAction
@@ -99,7 +99,7 @@ class PlayViewModelLikeTest {
                 likeInfo = likeModelBuilder.buildCompleteData(
                         param = likeModelBuilder.buildParam(),
                         status = likeModelBuilder.buildStatus(
-                                totalLike = initialTotalLike,
+                                totalLike = initialTotalLike.toLong(),
                                 totalLikeFormatted = initialTotalLike.toString(),
                                 isLiked = false
                         )
@@ -117,7 +117,7 @@ class PlayViewModelLikeTest {
                     .isAvailable()
                     .isEqualTo(
                             likeModelBuilder.buildStatus(
-                                    totalLike = newLikeCount,
+                                    totalLike = newLikeCount.toLong(),
                                     totalLikeFormatted = newLikeCount.toString(),
                                     isLiked = false,
                                     source = LikeSource.UserAction
